Reimplemented the NN code
Now splits the data without sklearn.
The model architecture is kept identical to the best performer.

Unfortunately, the author failed to set the tensorflow random seed.
This makes replicating the results kind of difficult.
If the data splitting turns out to be an issue it can be corrected for.

This reference implementation can be used as our springboard forwards.

It currently uses tensorboard for monitoring progress.
This should be switched for weights and biases at some point.