Implemented functions to extract molecular information.
Data is structured into a class MoleculeData.
Countable properties (ie number of atoms) are stored as Counters.
Features to be binned are stored as DefaultDict,
with keys as tuples of atomic numbers.

Some effort has gone to ensuring that the keys are always consistent in
their ordering to avoid any confusion; this is now explicit (closes #29 for good).