Added artifact load/save for parser/vectorizer
Implemented global filtering rules which always remove strucrures that
will break the vectorizer. Fragmented structures are always removed.
This is handled by the vectorizer, retaining the role of the parser as
an agnostic tool for converting molecules into an sdf archive.
Essentially it functions as an adapter.

Changed the convention of the vectorizer command to specify either input
with -i or artifact with -l flags. Allow for writing vectorizer to
artifact without saving locally with the -a flag.

Implemented the ML submodule to handle wandb tasks and splitting.
Removed the splitting submodule as its functionality is better placed in
the ml part of the library.

Completely reorganised MLP.py to make use of the new API.

Updated documentation to match.

Closes #75